test: change tox to pytest and keep integration test file name
