+++ tox.ini
@@ -12,7 +12,7 @@ install_command = pip install {opts} {packages}
 deps = -r{toxinidir}/requirements.txt
        -r{toxinidir}/test-requirements.txt
 commands =
-  python setup.py testr --testr-args='{posargs}'
+  pytest
 
 [testenv:pep8]
 commands =
